separate testing code into submodules
